% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pool.R
\name{db_release}
\alias{db_release}
\title{Return a connection checked out via \code{\link[=db_checkout]{db_checkout()}} to the pool}
\usage{
db_release(con)
}
\arguments{
\item{con}{A connection returned by \code{\link[=db_checkout]{db_checkout()}}.}
}
\description{
Return a connection checked out via \code{\link[=db_checkout]{db_checkout()}} to the pool
}
